move few components around, add drag handles, prepare code for dnd
