+++ jsapp/js/components/processing/analysis/analysisContent.module.scss
@@ -1,0 +1,5 @@
+@use './analysisVariables';
+
+.root {
+  padding: analysisVariables.$rows-gap 0;
+}
