--- jsapp/js/components/processing/analysis/analysisQuestionsList.module.scss
@@ -1,13 +1,0 @@
-@use 'scss/sizes';
-
-.root {
-  padding: sizes.$x20 0;
-}
-
-.list {
-  list-style: none;
-}
-
-.row:not(:last-child) {
-  margin-bottom: sizes.$x20;
-}
