+++ jsapp/js/components/processing/analysis/analysisVariables.scss
@@ -1,0 +1,6 @@
+@use 'scss/sizes';
+
+$icon-size: sizes.$x32;
+$rows-gap: sizes.$x20;
+$inner-vertical-gap: sizes.$x10;
+$inner-horizontal-gap: sizes.$x8;
