+++ jsapp/js/components/processing/analysis/list/analysisQuestionRow.module.scss
@@ -1,0 +1,34 @@
+@use 'scss/sizes';
+@use '~kobo-common/src/styles/colors';
+@use '../analysisVariables';
+
+.root {
+  display: flex;
+  flex-direction: row;
+}
+
+.content {
+  flex: 1;
+}
+
+.dragHandle {
+  color: colors.$kobo-gray-65;
+  width: analysisVariables.$icon-size;
+  line-height: analysisVariables.$icon-size;
+  cursor: move;
+  position: relative;
+
+  &:hover {
+    color: colors.$kobo-gray-40;
+  }
+
+  // We make the handle area bigger by adding this invisible element
+  &::after {
+    content: '';
+    width: calc(200% + analysisVariables.$icon-size);
+    height: calc(100% + analysisVariables.$rows-gap);
+    position: absolute;
+    top: analysisVariables.$rows-gap * -0.5;
+    left: calc(-50% - analysisVariables.$icon-size*0.5);
+  }
+}
