+++ jsapp/js/components/processing/analysis/list/analysisQuestionsList.module.scss
@@ -1,0 +1,11 @@
+@use 'scss/sizes';
+@use '~kobo-common/src/styles/colors';
+@use '../analysisVariables';
+
+.root {
+  list-style: none;
+
+  > *:not(:last-child) {
+    margin-bottom: analysisVariables.$rows-gap;
+  }
+}
